fix: handle auth links on footer EXLM-1195
